design for login/register pages
